+++ src/pages/HomePage.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const HomePage = () => <h1 className="title">Home page</h1>;
