+++ src/pages/NotFoundPage.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const NotFoundPage = () => <h1 className="title">Page not found</h1>;
