+++ src/pages/TabsPage.tsx
@@ -1,0 +1,35 @@
+import cn from 'classnames';
+import React from 'react';
+import { Link, useParams } from 'react-router-dom';
+const tabs = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
+
+export const TabsPage = () => {
+  const { tabId } = useParams();
+  const activeTab = tabs.find(t => t.id === tabId) || null;
+
+  return (
+    <>
+      <h1 className="title">Tabs page</h1>
+      <div className="tabs is-boxed">
+        <ul>
+          {tabs.map(t => (
+            <li
+              key={t.id}
+              data-cy="Tab"
+              className={cn({ 'is-active': t.id === activeTab?.id })}
+            >
+              <Link to={`/tabs/${t.id}`}>{t.title}</Link>
+            </li>
+          ))}
+        </ul>
+      </div>
+      <div className="block" data-cy="TabContent">
+        {activeTab ? activeTab.content : 'Please select a tab'}
+      </div>
+    </>
+  );
+};
